backend/requirements.txt
Django>=5.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
Pillow>=10.0

# API Documentation
drf-spectacular==0.27.0

# Testing
pytest==8.0.0
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.0
faker==22.0.0

# Security
django-ratelimit==4.1.0
